added animation to allproducts page but removed from userproducts page, fixed timer as at the same time only one product will be produced also used one timer, products and money calculations moved to initstate now decreasing money buying products is done, when remainingtime is zero amount, money, isProducing, starttime change in FB...adding product from allproducts page now changing green with rotation animation when finished returning orange with the animation...

+++ lib/models/users.dart
@@ -7,7 +7,7 @@ class Users {
   final String id;
   final String nickname;
   final String email;
-  int money = 0;
+  double money = 0;
   List<Map<String, Product>>  products = [];
   List<Factory> factories = [];
   List<City> cities = [];
